+++ test/devices_tests/sensor_test.py
@@ -775,6 +775,8 @@ class TestSensor:
         assert sensor.sensor_value.value == 16.96
         assert sensor.sensor_value.telegram.payload.value == DPTArray((0x06, 0xA0))
         assert sensor.resolve_state() == 16.96
+        # test HomeAssistant device class
+        assert sensor.ha_device_class() == "temperature"
 
     async def test_process_callback(self):
         """Test process / reading telegrams from telegram queue. Test if callback is called."""
@@ -793,11 +795,3 @@ class TestSensor:
         await sensor.process(telegram)
         after_update_callback.assert_called_with(sensor)
         assert sensor.last_telegram == telegram
-
-    def test_unique_id(self):
-        """Test unique id functionality."""
-        xknx = XKNX()
-        sensor = Sensor(
-            xknx, "TestSensor", group_address_state="1/2/3", value_type="temperature"
-        )
-        assert sensor.unique_id == "1/2/3"