+++ xknx/devices/__init__.py
@@ -11,6 +11,8 @@ from .expose_sensor import ExposeSensor
 from .fan import Fan
 from .light import Light
 from .notification import Notification
+from .numeric_value import NumericValue
+from .raw_value import RawValue
 from .scene import Scene
 from .sensor import Sensor
 from .switch import Switch
@@ -29,6 +31,8 @@ __all__ = [
     "Fan",
     "Light",
     "Notification",
+    "NumericValue",
+    "RawValue",
     "Scene",
     "Sensor",
     "Switch",