+++ xknx/devices/numeric_value.py
@@ -1,21 +1,17 @@
 """
-Module for managing a sensor via KNX.
+Module for managing a generic numeric value via KNX.
 
 It provides functionality for
 
-* reading the current state from KNX bus.
+* reading the current state from KNX bus or providing its state to the bus.
+* send local state changes to KNX bus.
 * watching for state updates from KNX bus.
 """
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, Iterator
+from typing import TYPE_CHECKING, Iterator
 
-from xknx.remote_value import (
-    GroupAddressesType,
-    RemoteValue,
-    RemoteValueControl,
-    RemoteValueSensor,
-)
+from xknx.remote_value import GroupAddressesType, RemoteValueNumeric
 
 from .device import Device, DeviceCallbackType
 
@@ -24,49 +20,34 @@ if TYPE_CHECKING:
     from xknx.xknx import XKNX
 
 
-class Sensor(Device):
-    """Class for managing a sensor."""
+class NumericValue(Device):
+    """Class for managing a numeric value."""
 
     def __init__(
         self,
         xknx: XKNX,
         name: str,
+        group_address: GroupAddressesType | None = None,
         group_address_state: GroupAddressesType | None = None,
+        respond_to_read: bool = False,
         sync_state: bool | int | float | str = True,
-        always_callback: bool = False,
         value_type: int | str | None = None,
         device_updated_cb: DeviceCallbackType | None = None,
     ):
         """Initialize Sensor class."""
         super().__init__(xknx, name, device_updated_cb)
+        self.respond_to_read = respond_to_read
+        self.sensor_value = RemoteValueNumeric(
+            xknx,
+            group_address=group_address,
+            group_address_state=group_address_state,
+            sync_state=sync_state,
+            value_type=value_type,
+            device_name=self.name,
+            after_update_cb=self.after_update,
+        )
 
-        self.sensor_value: RemoteValueControl | RemoteValueSensor
-        if isinstance(value_type, str) and value_type in [
-            "stepwise_dimming",
-            "stepwise_blinds",
-            "startstop_dimming",
-            "startstop_blinds",
-        ]:
-            self.sensor_value = RemoteValueControl(
-                xknx,
-                group_address_state=group_address_state,
-                sync_state=sync_state,
-                value_type=value_type,
-                device_name=self.name,
-                after_update_cb=self.after_update,
-            )
-        else:
-            self.sensor_value = RemoteValueSensor(
-                xknx,
-                group_address_state=group_address_state,
-                sync_state=sync_state,
-                value_type=value_type,
-                device_name=self.name,
-                after_update_cb=self.after_update,
-            )
-        self.always_callback = always_callback
-
-    def _iter_remote_values(self) -> Iterator[RemoteValue[Any, Any]]:
+    def _iter_remote_values(self) -> Iterator[RemoteValueNumeric]:
         """Iterate the devices RemoteValue classes."""
         yield self.sensor_value
 
@@ -77,11 +58,19 @@ class Sensor(Device):
 
     async def process_group_write(self, telegram: "Telegram") -> None:
         """Process incoming and outgoing GROUP WRITE telegram."""
-        await self.sensor_value.process(telegram, always_callback=self.always_callback)
-
-    async def process_group_response(self, telegram: "Telegram") -> None:
-        """Process incoming GroupValueResponse telegrams."""
         await self.sensor_value.process(telegram)
+
+    async def process_group_read(self, telegram: "Telegram") -> None:
+        """Process incoming GroupValueResponse telegrams."""
+        if (
+            self.respond_to_read
+            and telegram.destination_address == self.sensor_value.group_address
+        ):
+            await self.sensor_value.respond()
+
+    async def set(self, value: float | int) -> None:
+        """Set new value."""
+        await self.sensor_value.set(value)
 
     def unit_of_measurement(self) -> str | None:
         """Return the unit of measurement."""
@@ -91,13 +80,13 @@ class Sensor(Device):
         """Return the home assistant device class as string."""
         return self.sensor_value.ha_device_class
 
-    def resolve_state(self) -> Any | None:
+    def resolve_state(self) -> float | int | None:
         """Return the current state of the sensor as a human readable string."""
         return self.sensor_value.value
 
     def __str__(self) -> str:
         """Return object as readable string."""
-        return '<Sensor name="{}" sensor={} value={} unit="{}"/>'.format(
+        return '<NumericValue name="{}" addresses={} value={} unit="{}"/>'.format(
             self.name,
             self.sensor_value.group_addr_str(),
             self.resolve_state().__repr__(),