+++ xknx/remote_value/remote_value_datetime.py
@@ -34,13 +34,13 @@ class RemoteValueDateTime(RemoteValue[DPTArray, time.struct_time]):
         xknx: XKNX,
         group_address: GroupAddressesType | None = None,
         group_address_state: GroupAddressesType | None = None,
-        sync_state: bool = True,
+        sync_state: bool | int | float | str = True,
         value_type: str = "time",
         device_name: str | None = None,
         feature_name: str = "DateTime",
         after_update_cb: AsyncCallbackType | None = None,
     ):
-        """Initialize RemoteValueSensor class."""
+        """Initialize RemoteValueDateTime class."""
         try:
             self.dpt_class: type[DPTDate | DPTDateTime | DPTTime] = DateTimeType[
                 value_type.upper()